implemented authorization through policies for questions

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,14 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+	/**
+	 * Constructor
+	 * 
+	 * @return void
+	 */
+	public function __construct() {
+		$this->middleware('auth', ['except' => ['index', 'show']]);
+	}
 
 	/**
 	 * Display a listing of the resource.
@@ -73,6 +81,7 @@ class QuestionsController extends Controller
 	 */
 	public function edit(Question $question)
 	{
+		$this->authorize('update', $question);
 		return view('questions.edit')->with('question', $question);
 	}
 
@@ -85,6 +94,7 @@ class QuestionsController extends Controller
 	 */
 	public function update(AskQuestionRequest $request, Question $question)
 	{
+		$this->authorize('update', $question);
 		if ($request->body != $question->body || $request->title != $question->title) {
 			$question->update($request->only('title', 'body'));
 			return redirect()->route('questions.index')->with('success', __('Your question has been updated'));
@@ -99,6 +109,7 @@ class QuestionsController extends Controller
 	 */
 	public function destroy(Question $question)
 	{
+		$this->authorize('delete', $question);
 		$question->delete();
 		return redirect()->route('questions.index')->with('success', __('Your question has been deleted'));
 	}
